app/requirements.txt
dash
dash-bootstrap-components
pandas
numpy
opencv-python
Pillow
dash_extensions
waitress